+++ CHANGELOG.md
@@ -84,8 +84,8 @@
 - API: add PrintColumns for KongTarget upstream and target fields.
   [#4576](https://github.com/Kong/kong-operator/pull/4576)
 - AIGateway:
-  - introduce `AIGatewayDataPlane` CRD
-  [#4690](https://github.com/Kong/kong-operator/pull/4690)
+  - introduce `AIGatewayDataPlane` CRD with reconciler
+  [#4690](https://github.com/Kong/kong-operator/pull/4690), [#4797](https://github.com/Kong/kong-operator/pull/4797)
   - introduce `AIGatewayControlPlane` CRD with reconciler
   [#4760](https://github.com/Kong/kong-operator/pull/4760), [#4792](https://github.com/Kong/kong-operator/pull/4792)
   - introduce `AIGatewayDataPlaneCertificate` CRD with reconciler